// frontend/src/types/bot.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Bot = { id: bigint, owner_id: bigint | null, token: string, username: string, type: BotType, is_active: boolean, is_primary: boolean, referral_percentage: number, created_at: string, updated_at: string, created_by: bigint | null, };

export type BotType = "main" | "referral";

export type NewBot = { token: string, };

export type UpdateBot = { is_active?: boolean, is_primary?: boolean, referral_percentage?: number, };